Update designing.html.erb.md
Reopened because the state indication had become a little off - it is not an example but an additional activity. Also, it is not only for links but also for buttons and other interactive elements.

I still miss the word "controls" in there, though. Is "interactive elements" technically more precise (ie. "links" may not be perceived as "controls")? - but will people understand that custom widgets are also in scope of this tip?